acet-gui first commit
Basically I was toying around with wxWidgets and making a gui version
last year. This is basically what I got up to. Barely tested and
probably hilariously broken so if anyone wants to try it, enter at your
own risk. I'm gonna try and work on it this year but I can't
promise anything. Surely someone probably has a better emblem tool out
by now but I just wanna do it for funsies and to practice some ui/gui
stuff. 'acet-cli' is basically the coke classic version of the tool.
cmake seems to build it fine on my machine but I'm pretty bad at cmake
so I won't be surprised if there are bugs in the build process. Happy
new year!