Attempt at fixing images on injection
Added image scaling and color quantizing. Not really tested in game at all just
committing progress. wxQuantize was giving me a headache for a while
as it would just crash the program at different times. Figured it out and got
it mostly working on the couple of images I was testing with. Certain there are
bugs with it that I'll get to later.

Also accidentally had clang formatter on and saved acet-gui.cpp so it kinda fucked
my formatting and I had to redo/respace things