FindPaletteOffset and checking dropped saves
FindPaletteOffset looks for the first two entries in the palette map
( 0x00 0x00 0x00 0x00 and * * * 0x80). Probably a better method of finding
the emblem data in a save than what we have so I'll have to get around to
using it for the actual extract and inject logic. For now its just being used
to verify that the save files dropped are actual saves. Doesn't check if it is
an emblem save vs single player save but I think I wanna make the tool work on
single player saves too (every single player save has the whole darn emblem in
it - I think the tricky part might be the different file endings).